programming model-v2/ai-search-samples/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai==1.2.*
numpy
azure-search-documents==11.4.0
